First try at fixing issue #26.
Callback members of structs are now read XOR write.
i.e.  They can be set in java for native code to callback, or they can be set by native code and called by java, but not both.

Not sure if re-using callback for this was the best path, or if a new interface (Callable? Invokable?) should have been used just for this.

git-svn-id: https://svn.java.net/svn/jna~svn/branches/jnalib-v3@206 2f8a963e-d2e4-e7d0-97bf-ccb7fcea9d80